Pin local to production version (more below)
This is a reversal of opinion on my part. I imagined that we would docker pull the graphql-engine image in our local
stack more than I think we do. This will keep our local stacks
up to date, and everyone is free to overload this in their
checkouts if they want.